+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [2.1.0](https://github.com/yoch/modpoll2mqtt/compare/v2.0.0...v2.1.0) (2026-06-10)
 
 ### Features